requirements.txt
requests
feedparser
google-genai
python-dotenv
beautifulsoup4==4.14.3
python-dateutil==2.9.0.post0
tenacity==9.1.4
